worker/requirements.txt
temporalio==1.9.0
chromadb==0.6.3
httpx==0.28.1
boto3==1.36.0
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5
